Make the ImagePullPolicy Configurable (#237)
The pull policy of an APB should be set by the broker.
To configure it, set image_pull_policy in the config file.

The image_pull_policy setting will be verified that it's
a correct setting. The option is not case sensetive and will
error if an incorrect setting is used.