initial attempt at multiple registries.
Adding ability to catch ID collison and to warn user which spec is overwritting.

fix build issues when refactoring registry to its own package.

Adding registry failure bahavior

* if all fail, fail the bootstrap
* if 1 fails and is marked on fail_on_error then fail the bootstrap\
  this does not continue the bootstrap
* Adding method that will take an error and determine if the registry should fail

refactor so that the dockerhub and rhcc are using the same image data -> spec.

* stop pulling down entire image for dockerhub
* using regisitry.dockerhub.com to get the manifest which is the same manifest as RHCC
* able to combine image data -> spec.

Multiple registries that are using the same manifest EP.

* Allows us to generate the spec ID.
* attempts to make the service name unique by keeping the registry with it.

fixing bug with name not valid.