Improve user facing documentation for broker
Use openshift/origin and kubernetes/kubernetes as an example of what the
project's root README.md should look like. Update the CONTRIBUTING guide
with steps for setting up the environment and for what to do before
creating a PR. Also, rename `pre-flight` make target to `check` as that
makes more sense as the PR "gate".

Added OWNERS file to call out reviewers/approvers.