Use origin-ansible-service-broker docker image (#371)
* Simplify run latest build script

It is annoying for the run_latest_build script to cat the broker k8s
object to a file and then use `oc create` when it could just pipe
directly.

* Add origin- prefix to ansible-service-broker image

A majority of the work to make the change from
ansible-service-broker to origin-ansible-service-broker is taken care of
through jenkins pipelines and dockerhub automated builds. Now that those
changes are in place, this updates the relevant documentation + the
broker template that we use to reference the correct image
(`origin-ansible-service-broker`).