Explain how to run a locally built broker
Expose new contributors to the broker configuration file as early as
possible by pointing them towards it when they go to run the broker
locally.

Update the local_dev_vars example to use changeme/changeme for dockerhub
as this won't fail if you just copy/paste and run.

Update the prep_local_devel_env script to cover both catasb and
run_latest_build case. The way the specs are stored in etcd is why one
would need to be careful with the registry name.