APB Sandbox Role should be configurable
As mentioned in #399, there are apb's no and in the future that will
wish to create and/or modify service accounts. Until such a time as
apb's asking for the permissions (roles) that it will need, this moves
closer to that goal by making it a configurable item in the cluster
config section of the broker.

- Update example's and script's that generate a broker config to include
  a sandbox role (default: edit)
- Update apb package to pass roll on creation of sandbox

Fixes #399