Create the broker endpoint separately
If you create endpoint 'asb', it will be associated with
the service 'asb'. That service, doesn't connect to a running
pod so the endpoint gets no ip address. Instead, create another
endpoint that connect to the service on the same ip:port to use
as the endpoint to contact the broker.