Make the prep_local_devel_env script work for Kubernetes & Openshift (#434)
* Make the prep_local_devel_env script work for Kubernetes & Openshift

Most of the prep_local_devel_env.sh script was shared between the
two clusters. The main differences were:
  - etcd_port: 2379
  - kubernetes_port: 6443
  - routes
  - deployments
  - ca.crt vs service-ca.crt

* Create the broker endpoint separately

If you create endpoint 'asb', it will be associated with
the service 'asb'. That service, doesn't connect to a running
pod so the endpoint gets no ip address. Instead, create another
endpoint that connect to the service on the same ip:port to use
as the endpoint to contact the broker.

* Reorganize scirpts

* Now that we create asb-1338, make sure we delete it

* Never commited the k8s-local-dev-changes template

* Remove empty script

* Remove cluster-redirect script