Integrate a new ci framework for travis (#463)
* Test out local ci

* Use the latest broker-ci repo instead of a release for now

* Carry the exit code through travis.sh

* Install the go pkg with go get and carry a config.yaml

This repo should carry it's own config.yaml since it's a consumer
of the new framework.

Remove the local ci scripts and move the templates to the templates
directory.

* Still need error.sh, logs.sh, and wait-for-resource.sh

* A few improvements to .travis

- Add folded up broker an catalog logs to the end of a run
- Error properly when setup.sh fails

* Change the verification test

Verify the secret is created instead of the podpreset

* Some debugging changes

* Add some debugging to config.yaml

* Only print the logs we need and further debugging

* Cleanup the broker scripts

We don't need the broker scripts anymore. We're moving to the new
framework.

* Few changes to config.yaml

* Split up the cluster deployment and broker deloyment

Make these scripts individually consumable so the APB gates can
use them.

* Update config.yaml to grab framework 3.7 changes

* Update gather-logs.sh to 3.7 resources

* Make sure the deployment name is the same as the serviceinstance name

* Update the gather-logs.sh script

* Make sure non broker CI jobs set up the broker with catasb

The gate job for the broker runs the asb with shell scripts.
Make sure we use catasb for non asb jobs.