Broker should extract credentials from secret (#555)
* Broker should extract credentials from secret

This change makes it so the broker can handle secrets that are created
by APBs when using the `asb_encode_binding` module.

- Update the broker so that it can handle secrets generated by the APB
  when `asb_encode_binding` module is used from the asb-modules.
- Update `executor::ExecuteApb` to wait for pod to complete, since the
  pod is no longer kept alive for credential extraction.
- Clean up some of the log messages and code format related to apb
  actions.

Fixes #544
Fixes #553
Implements the proposal #550

Depends on the following PRs:
- [ansible-asb-modules#8](ansibleplaybookbundle/ansible-asb-modules#8) This is how the secret gets generated.
- [apb-base#7](ansibleplaybookbundle/apb-base#7) Remove scripts related to extracting credentials from the containers filesystem.
- [ansible-playbook-bundle#163](ansibleplaybookbundle/ansible-playbook-bundle#163) Bump the APB versions so freshly built APBs will pass version validation checks.

* Add APB runtime version to our APB Spec object

Update the broker to evaluate the `com.redhat.apb.runtime` label on APBs
(default to `1` when there is no label). Add version checking of this
new min/max apb runtime version and update associated tests.

* Handle bind credential extraction based on runtime

Update apb execution to be handle multiple apb runtime versions.

* Fixing log statements

* Only decode bind creds if encoded

Secrets retrieved using client-go are already decoded. So the extract
credentials function has been updated to only attempt to decode the
credentials if they need it.

Also changed the `log.Error` when the APB completed to a `log.Notice`
since it is not a failure.

* Pull k8s API call to get pod status into k8s client

* Cleanup extract credentials based on comments

* Update proposal based on what is implemented

* Improve adapter log info on apb runtime eval

* Fixes after rebase

* Address review comments and fix unit tests

* Ask for the k8s client when we need it

* Use canary APB images