fix statements with formatting (#904)
* fix statements with formatting

Making broker consistent with the log statements in bundle-lib, i.e. use
the *f version when logging formatting strings. It is most important
when using logrus, go-logging seems to handle it okay. But this PR will
make it easier to move to logrus in the future.

See PR #61 from bundle-lib for more information:

https://github.com/automationbroker/bundle-lib/pull/61

* revert goimports auto imports