Bug 1583064 - missing action causes many sandboxes (#972)
* Bug 1583064 - missing action causes many sandboxes

If the unbind fails we really want to let them know it failed because it
is probably something the APB developer needs to fix. But if the action
is non-existent we really need to prevent the multiple sandboxes.

If we can't find the action, there is no point in continuing.
marking job as SUCCEEDED but setting the message to be the actual error
string. This will only happen for the optional actions: bind & unbind.

Update isJobInProgress to take a string to allow the proper ID to be
passed in.

* remove the now useless debug statement