Bug 1615723 - check for not found ext creds (#1051)
A regression was introduced that caused us not to handle the case when
the extracted credentials were not found in the switch statement. It
used to work because we were doing the client.IsNotFoundError for both
checks. We switched one to checking for an exact error:
ErrExtractedCredentialsNotFound but didn't update the switch statement.